feat: implement update advertisement use case + changes to related classes

+++ redline-core/src/main/java/com/yuriolivs/redlinecore/usecase/advertisement/UpdateAdUseCase.java
@@ -20,7 +20,7 @@ public class UpdateAdUseCase {
 
     public Advertisement execute(
             String url,
-            BigDecimal price
+            Double price
     ) {
         throw new UnsupportedOperationException("Not implemented yet.");
     }
